+++ examples-next/basic/schema.ts
@@ -146,13 +146,6 @@ export const lists = createSchema({
             labelField: 'name',
             listKey: 'User',
           },
-          manyBlockMentions: {
-            kind: 'block',
-            label: 'Many mentions',
-            labelField: 'name',
-            listKey: 'User',
-            many: true,
-          },
           featuredAuthors: {
             kind: 'prop',
             labelField: 'name',
@@ -233,7 +226,7 @@ export const extendGraphqlSchema = graphQLSchemaExtension({
       },
     },
     Mutation: {
-      createRandomPosts(root: any, args: any, context: any) {
+      createRandomPosts(root, args, context) {
         // TODO: add a way to verify access control here, e.g
         // await context.verifyAccessControl(userIsAdmin);
         const lists: KeystoneListsAPI<KeystoneListsTypeInfo> = context.lists;