+++ packages-next/auth/src/gql/getBaseAuthSchema.ts
@@ -1,3 +1,5 @@
+import type { GraphQLSchemaExtension } from '@keystone-next/types';
+
 import { AuthGqlNames } from '../types';
 
 import { validateSecret } from '../lib/validateSecret';
@@ -15,7 +17,7 @@ export function getBaseAuthSchema({
   secretField: string;
   protectIdentities: boolean;
   gqlNames: AuthGqlNames;
-}) {
+}): GraphQLSchemaExtension {
   return {
     typeDefs: `
       # Auth
@@ -46,7 +48,11 @@ export function getBaseAuthSchema({
     `,
     resolvers: {
       Mutation: {
-        async [gqlNames.authenticateItemWithPassword](root: any, args: any, context: any) {
+        async [gqlNames.authenticateItemWithPassword](root, args, context) {
+          if (!context.startSession) {
+            throw new Error('No session implementation available on context');
+          }
+
           const list = context.keystone.lists[listKey];
           const itemAPI = context.lists[listKey];
           const result = await validateSecret(
@@ -76,7 +82,7 @@ export function getBaseAuthSchema({
         },
       },
       Query: {
-        async authenticatedItem(root: any, args: any, { session, lists }: any) {
+        async authenticatedItem(root, args, { session, lists }) {
           if (typeof session?.itemId === 'string' && typeof session.listKey === 'string') {
             const item = await lists[session.listKey].findOne({ where: { id: session.itemId } });
             return item || null;
@@ -85,7 +91,7 @@ export function getBaseAuthSchema({
         },
       },
       AuthenticatedItem: {
-        __resolveType(rootVal: any, { session }: any) {
+        __resolveType(rootVal, { session }) {
           return session?.listKey;
         },
       },