+++ packages/keystone/tests/Keystone.test.js
@@ -116,6 +116,20 @@ describe('Keystone.createList()', () => {
     expect(keystone.listsArray[0]).toBe(keystone.lists['User']);
   });
 
+  test('Reserved words', () => {
+    const config = {
+      adapter: new MockAdapter(),
+      cookieSecret: 'secretForTesting',
+    };
+    const keystone = new Keystone(config);
+
+    for (const listName of ['Query', 'Mutation', 'Subscription']) {
+      expect(() => keystone.createList(listName, {})).toThrow(
+        `Invalid list name "${listName}". List names cannot be reserved GraphQL keywords`
+      );
+    }
+  });
+
   /* eslint-disable jest/no-disabled-tests */
   describe('access control config', () => {
     test.todo('expands shorthand acl config');