+++ examples-next/basic/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @keystone-next/app-basic
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [[`b6498d9f1`](https://github.com/keystonejs/keystone/commit/b6498d9f1341648742f2db78fec53b851b36dddd), [`dfed54074`](https://github.com/keystonejs/keystone/commit/dfed5407477ffa387b38bf76685297941f5d8f84), [`d36e580cc`](https://github.com/keystonejs/keystone/commit/d36e580cc21e4b77a1bd0615c96c0793b9c5dac5), [`803626e88`](https://github.com/keystonejs/keystone/commit/803626e8854f9b7d293bd1829398d25a6692154a)]:
+  - @keystone-next/auth@6.0.0
+  - @keystone-next/keystone@5.0.0
+  - @keystone-next/types@5.0.0
+  - @keystone-next/admin-ui@3.1.2
+  - @keystone-next/fields@3.2.1
+  - @keystone-next/fields-document@0.0.2
+
+## 0.0.6
+
+### Patch Changes
+
+- [`443c45934`](https://github.com/keystonejs/keystone/commit/443c45934ba9adae3a33c95f1d567b84568006ad) [#4403](https://github.com/keystonejs/keystone/pull/4403) Thanks [@jossmac](https://github.com/jossmac)! - test document-field primitives
+
+* [`55a04a100`](https://github.com/keystonejs/keystone/commit/55a04a1004b7f15fcd41b4935d74fd1847c9deeb) [#4399](https://github.com/keystonejs/keystone/pull/4399) Thanks [@jossmac](https://github.com/jossmac)! - Document field component block UI
+
+- [`f57a2e136`](https://github.com/keystonejs/keystone/commit/f57a2e1363373c33a2718b0b3c3a07f3770513ac) [#4416](https://github.com/keystonejs/keystone/pull/4416) Thanks [@jossmac](https://github.com/jossmac)! - add custom "notice" block to document field
+
+* [`6cd469e29`](https://github.com/keystonejs/keystone/commit/6cd469e29682ff41a515ed76919efc2bfe0c7567) [#4414](https://github.com/keystonejs/keystone/pull/4414) Thanks [@JedWatson](https://github.com/JedWatson)! - Typed keystone context
+
+- [`139c74a4a`](https://github.com/keystonejs/keystone/commit/139c74a4aacfd5230fc86471b9cbd2a3c90bac9a) [#4378](https://github.com/keystonejs/keystone/pull/4378) Thanks [@timleslie](https://github.com/timleslie)! - Updated code to consistently use `context` rather than `ctx` for graphQL context variables.
+
+- Updated dependencies [[`8543f732b`](https://github.com/keystonejs/keystone/commit/8543f732bfc33c193675c1db385b8e2c15d3f273), [`8b12f795d`](https://github.com/keystonejs/keystone/commit/8b12f795d64dc085ca663921aa6826350d234cd0), [`2c5135f17`](https://github.com/keystonejs/keystone/commit/2c5135f17ccf94412f074b1bdd04592ea7c25d31), [`8baad4a14`](https://github.com/keystonejs/keystone/commit/8baad4a147e224bf99daeb91337d2eedea2f879b), [`55a04a100`](https://github.com/keystonejs/keystone/commit/55a04a1004b7f15fcd41b4935d74fd1847c9deeb), [`2926369a0`](https://github.com/keystonejs/keystone/commit/2926369a00134f361c70b046aa2b9cf7674740f7), [`1cf445cad`](https://github.com/keystonejs/keystone/commit/1cf445cad1017b2ad8b7337649783e75ba4fe7d4), [`f4a855c71`](https://github.com/keystonejs/keystone/commit/f4a855c71e966ef3ebc894a3b0f1af51e5182394), [`c9159c956`](https://github.com/keystonejs/keystone/commit/c9159c9560efa7f7ae6ef802302c97b7a23f987b), [`6cd469e29`](https://github.com/keystonejs/keystone/commit/6cd469e29682ff41a515ed76919efc2bfe0c7567), [`fa12a18b0`](https://github.com/keystonejs/keystone/commit/fa12a18b077367563b1b69db55274e47a1bd5027), [`0082a3144`](https://github.com/keystonejs/keystone/commit/0082a3144b93850bf4841a8e09f91f43b303088e), [`eddd7e795`](https://github.com/keystonejs/keystone/commit/eddd7e79599e20f5bef61f240d874b37d0a084c4), [`bb747c9cb`](https://github.com/keystonejs/keystone/commit/bb747c9cbe4ebb56eecc70be2e90f77165e5fab6), [`139c74a4a`](https://github.com/keystonejs/keystone/commit/139c74a4aacfd5230fc86471b9cbd2a3c90bac9a), [`2ab785c79`](https://github.com/keystonejs/keystone/commit/2ab785c793231555ac254b24908ca0e8d1ac7c95), [`b8c2c48ec`](https://github.com/keystonejs/keystone/commit/b8c2c48ec3746809894af7347c205f6a95329e8d), [`2dccc86cf`](https://github.com/keystonejs/keystone/commit/2dccc86cfbdbd1b11f3ad63831e1b08d66a12342)]:
+  - @keystone-next/auth@5.0.1
+  - @keystone-next/fields@3.2.0
+  - @keystone-ui/core@1.0.3
+  - @keystone-next/admin-ui@3.1.1
+  - @keystone-next/keystone@4.1.1
+  - @keystone-next/types@4.1.1
+  - @keystone-ui/tooltip@1.0.3
+  - @keystone-next/fields-document@0.0.1
+
 ## 0.0.5
 
 ### Patch Changes